port(zle): tc_downcurs faithful — tcmultout + newline fallback (c:2320-2331)
The Rust tc_downcurs faked the move as an unconditional CSI B and was
mis-signatured (void vs C's int). A plain CSI B moves within existing
content; it cannot scroll/create lines past what's drawn. Port the real
body: try tcmultout(TCDOWN, TCMULTDOWN, ct), and when no capability
exists emit ct real newlines + CR (which do grow the display) and return
-1 so callers know the column reset. Returns 0 on the capability path.
Update the two complist.rs callers to the i32 signature (return ignored,
as in C). Proof: tc_downcurs_newline_fallback_and_capability checks both
the newline fallback (-1, 3 newlines + CR) and the capability path (0,
looped escape). Fixed a self-deadlock in the test's tclen snapshot.

+++ src/ported/zle/complist.rs
@@ -2085,7 +2085,7 @@ pub fn singledraw() -> i32 {
 
     if md1 != 0 {
         // c:1952
-        tc_downcurs(md1 as usize); // c:1953
+        tc_downcurs(md1); // c:1953
     }
     if mc1 != 0 {
         // c:1954
@@ -2130,7 +2130,7 @@ pub fn singledraw() -> i32 {
     // c:1964-1965 — relative down-move to second cell.
     if md2 != md1 {
         // c:1964
-        tc_downcurs((md2 - md1) as usize); // c:1965
+        tc_downcurs(md2 - md1); // c:1965
     }
     if mc2 != 0 {
         // c:1966
